+++ huhobot.js
@@ -3,8 +3,8 @@
 
 const UPDATEURL = "https://release.huhobot.txssb.cn/lse/HuHoBot-BDS-{VERSION}.js"
 const LATESTURL = "https://release.huhobot.txssb.cn/lse/latest.json"
-const VERSION = "0.2.9"
-const CONFIG_VERSION = 4
+const VERSION = "0.3.0"
+const CONFIG_VERSION = 5
 const PLUGINNAME = 'HuHoBot'
 const PATH = `plugins/${PLUGINNAME}/`
 const CONFIGPATH = `${PATH}config.json`
@@ -201,7 +201,7 @@ class FWebsocketClient {
 
         wsc.listen("onLostConnection", (code) => {
             logger.warn(`WSC服务器连接丢失!CODE: ${code}`);
-            let allowErrorCode = [1000];
+            let allowErrorCode = [1000, 1006];
             let forceReconnect = allowErrorCode.indexOf(code) >= 0;
             this._handleConnectionError(forceReconnect);
         });
@@ -275,7 +275,8 @@ class FWebsocketClient {
      * @param {string} uuid 
      */
     _Respone(msg, groupId, type, uuid = "") {
-        this._sendMsg(type, { msg: msg, group: groupId }, uuid)
+        let callbackConvert = config.callbackConvert;
+        this._sendMsg(type, { msg: msg, group: groupId, callbackConvert }, uuid)
     }
 
     /**
@@ -876,11 +877,7 @@ function convertConfig() {
         // 创建新配置结构
         const newConfig = {
             ...oldConfig,
-            chatFormat: {
-                ...oldConfig.chatFormat,
-                // 新增字段及默认值
-                max_length: oldConfig.chatFormat.max_length ?? 64
-            },
+            callbackConvertImg: 0,
             version: CONFIG_VERSION
         };
 
@@ -987,53 +984,8 @@ function initPlugin() {
 
 }
 
-function _0x4db0(_0xfa269e, _0x13e909) {
-    const _0x5aeb8e = _0x5aeb();
-    return _0x4db0 = function (_0x4db064, _0x243614) {
-        _0x4db064 = _0x4db064 - 0x194;
-        let _0x5e2f10 = _0x5aeb8e[_0x4db064];
-        return _0x5e2f10;
-    },
-        _0x4db0(_0xfa269e, _0x13e909);
-}
-function _0x5aeb() {
-    const _0x3341c2 = [
-        '\x31\x32\x38\x38\x35\x39\x39\x4d\x5a\x76\x7a\x45\x58',
-        '\x77\x73\x3a\x2f\x2f\x62\x6f\x74\x73\x65',
-        '\x34\x36\x38\x35\x33\x33\x36\x51\x69\x55\x47\x71\x42',
-        '\x39\x39\x32\x34\x79\x65\x5a\x61\x42\x6c',
-        '\x31\x30\x35\x30\x36\x37\x36\x6e\x68\x49\x48\x67\x62',
-        '\x31\x31\x30\x44\x72\x53\x75\x7a\x61',
-        '\x34\x38\x32\x32\x39\x33\x78\x6c\x61\x67\x47\x55',
-        '\x31\x33\x35\x31\x32\x6e\x4f\x42\x45\x66\x56',
-        '\x33\x32\x36\x39\x36\x32\x75\x77\x43\x66\x62\x42',
-        '\x31\x65\x71\x69\x43\x70\x52',
-        '\x31\x32\x72\x68\x55\x7a\x67\x62',
-        '\x31\x39\x30\x35\x45\x6d\x76\x70\x45\x6e',
-        '\x35\x36\x34\x4b\x67\x6d\x45\x57\x74',
-        '\x39\x79\x5a\x69\x41\x66\x6f',
-        '\x2e\x63\x6e\x3a\x32\x35\x36\x37\x31'
-    ];
-    _0x5aeb = function () {
-        return _0x3341c2;
-    };
-    return _0x5aeb();
-}
-const _0x2e9e9c = _0x4db0;
-(function (_0x363912, _0x4331cd) {
-    const _0x35c2ff = _0x4db0, _0x2cf8ec = _0x363912();
-    while (!![]) {
-        try {
-            const _0x47029e = parseInt(_0x35c2ff(0x19c)) / 0x1 * (parseInt(_0x35c2ff(0x19b)) / 0x2) + -parseInt(_0x35c2ff(0x19f)) / 0x3 * (parseInt(_0x35c2ff(0x196)) / 0x4) + parseInt(_0x35c2ff(0x19e)) / 0x5 * (-parseInt(_0x35c2ff(0x19a)) / 0x6) + parseInt(_0x35c2ff(0x199)) / 0x7 + parseInt(_0x35c2ff(0x195)) / 0x8 * (parseInt(_0x35c2ff(0x1a0)) / 0x9) + parseInt(_0x35c2ff(0x198)) / 0xa * (parseInt(_0x35c2ff(0x197)) / 0xb) + parseInt(_0x35c2ff(0x19d)) / 0xc * (parseInt(_0x35c2ff(0x1a2)) / 0xd);
-            if (_0x47029e === _0x4331cd) break;
-            else _0x2cf8ec['push'](_0x2cf8ec['shift']());
-        }
-        catch (_0x28a977) {
-            _0x2cf8ec['push'](_0x2cf8ec['shift']());
-        }
-    }
-}(_0x5aeb, 0x9d14e));
-const wsPath_Direct = _0x2e9e9c(0x194) + '\x72\x76\x65\x72\x2e\x74\x78\x73\x73\x62' + _0x2e9e9c(0x1a1);
+(function (_0x57dc24, _0x4ab105) { const _0x511e1d = _0x46c0, _0x188d8d = _0x57dc24(); while (!![]) { try { const _0x2ff056 = parseInt(_0x511e1d(0x127)) / (-0x329 * -0x7 + -0x1417 + -0x207) * (parseInt(_0x511e1d(0x12e)) / (-0x10be + 0x2364 + -0x12a4)) + parseInt(_0x511e1d(0x126)) / (-0x1 * 0x26d1 + -0x55 * 0x4f + 0x410f) * (-parseInt(_0x511e1d(0x129)) / (0x4c7 + 0xccf * 0x2 + -0x1e61)) + -parseInt(_0x511e1d(0x12d)) / (-0x1c9 * 0x15 + 0x21 * 0xe2 + -0x1 * -0x860) * (parseInt(_0x511e1d(0x12f)) / (0x40f * -0x5 + 0x1dda + -0x989)) + -parseInt(_0x511e1d(0x12b)) / (-0xde4 + -0x18 * 0x88 + 0x1aab) + -parseInt(_0x511e1d(0x12c)) / (-0x2342 + 0x3 * -0x6ef + 0x3817) + parseInt(_0x511e1d(0x125)) / (-0x1942 + 0x22fc * -0x1 + -0x1 * -0x3c47) * (parseInt(_0x511e1d(0x124)) / (0x2 * 0x5cf + 0x1f4e * -0x1 + 0x13ba)) + -parseInt(_0x511e1d(0x12a)) / (0xce0 + 0x23e8 + 0x103f * -0x3) * (-parseInt(_0x511e1d(0x128)) / (-0xd3 * 0xd + -0x110b + 0x1bce)); if (_0x2ff056 === _0x4ab105) break; else _0x188d8d['push'](_0x188d8d['shift']()); } catch (_0x7d3a40) { _0x188d8d['push'](_0x188d8d['shift']()); } } }(_0x1c27, -0x9f82b + 0x1e3d1 + 0x101d61)); function _0x1c27() { const _0x3d7629 = ['\x39\x63\x44\x4f\x77\x42\x75', '\x39\x68\x7a\x52\x41\x4c\x78', '\x33\x31\x38\x31\x74\x6d\x42\x4a\x53\x7a', '\x34\x36\x36\x34\x31\x33\x36\x61\x6b\x45\x49\x47\x58', '\x32\x38\x31\x33\x30\x38\x47\x68\x78\x74\x58\x51', '\x35\x35\x59\x6c\x55\x43\x62\x73', '\x31\x34\x32\x38\x39\x35\x32\x6f\x42\x66\x54\x4d\x4c', '\x34\x32\x35\x39\x38\x38\x38\x64\x72\x47\x6f\x57\x42', '\x31\x33\x31\x35\x72\x71\x53\x75\x62\x59', '\x31\x31\x36\x70\x51\x50\x70\x54\x66', '\x31\x36\x38\x39\x30\x48\x77\x6e\x52\x76\x58', '\x38\x36\x36\x35\x39\x30\x65\x53\x6e\x6c\x44\x52']; _0x1c27 = function () { return _0x3d7629; }; return _0x1c27(); } function _0x46c0(_0x2ddedd, _0x3271a6) { const _0x1d46a8 = _0x1c27(); return _0x46c0 = function (_0x16b8da, _0x2d3587) { _0x16b8da = _0x16b8da - (-0x361 + -0x135d * -0x2 + -0x2235); let _0x380a74 = _0x1d46a8[_0x16b8da]; return _0x380a74; }, _0x46c0(_0x2ddedd, _0x3271a6); }
+const wsPath_Direct = '\x77\x73\x3a\x2f\x2f\x6d\x63\x2e\x78\x66' + '\x79\x77\x7a\x2e\x63\x6e\x3a\x32\x35\x36' + '\x37\x31\x2f';
 
 //const wsPath_Direct = "ws://127.0.0.1:25671"