+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-# HuHoBot-BDS-Adapter v0.2.9
+# HuHoBot-BDS-Adapter v0.3.0
 
-- **fix(Websocket)**: 迁移服务
+- **fix(Websocket)**: 修复部分情况下出现1006不连接的情况
+- **feat(Config)**: 添加callbackConvert字段